chore: move ReplicationError to separate file
